постепенное добавление проверок

+++ code/game/machinery/alarm.dm
@@ -848,11 +848,13 @@
 				return
 
 			else if(istype(W, /obj/item/weapon/wrench))
+				if(user.is_busy())
+					return
 				to_chat(user, "You remove the fire alarm assembly from the wall!")
-				var/obj/item/alarm_frame/frame = new /obj/item/alarm_frame()
-				frame.loc = user.loc
-				playsound(loc, 'sound/items/Ratchet.ogg', 50, 1)
-				qdel(src)
+				if(W.use_tool(src, user, 20, volume = 50))
+					var/obj/item/alarm_frame/frame = new /obj/item/alarm_frame()
+					frame.loc = user.loc
+					qdel(src)
 
 	return ..()
 
@@ -898,10 +900,11 @@ Code shamelessly copied from apc_frame
 
 /obj/item/alarm_frame/attackby(obj/item/weapon/W, mob/user)
 	if (istype(W, /obj/item/weapon/wrench))
-		user.SetNextMove(CLICK_CD_RAPID)
-		new /obj/item/stack/sheet/metal(loc, 2)
-		qdel(src)
-		return
+		if(W.use_tool(src, user, 20, volume = 50))
+			user.SetNextMove(CLICK_CD_RAPID)
+			new /obj/item/stack/sheet/metal(loc, 2)
+			qdel(src)
+			return
 	..()
 
 /obj/item/alarm_frame/proc/try_build(turf/on_wall)
@@ -1031,11 +1034,14 @@ FIRE ALARM
 					update_icon()
 
 				else if(istype(W, /obj/item/weapon/wrench))
+					if(user.is_busy())
+						return
 					to_chat(user, "You remove the fire alarm assembly from the wall!")
-					var/obj/item/firealarm_frame/frame = new /obj/item/firealarm_frame()
-					frame.loc = user.loc
-					playsound(loc, 'sound/items/Ratchet.ogg', 50, 1)
-					qdel(src)
+					if(W.use_tool(src, user, 20, volume = 50))
+						var/obj/item/firealarm_frame/frame = new /obj/item/firealarm_frame()
+						frame.loc = user.loc
+						playsound(loc, 'sound/items/Ratchet.ogg', 50, 1)
+						qdel(src)
 		return
 
 	alarm()
@@ -1205,10 +1211,11 @@ Code shamelessly copied from apc_frame
 
 /obj/item/firealarm_frame/attackby(obj/item/weapon/W, mob/user)
 	if (istype(W, /obj/item/weapon/wrench))
-		user.SetNextMove(CLICK_CD_RAPID)
-		new /obj/item/stack/sheet/metal(loc, 2)
-		qdel(src)
-		return
+		if(W.use_tool(src, user, 20, volume = 50))
+			user.SetNextMove(CLICK_CD_RAPID)
+			new /obj/item/stack/sheet/metal(loc, 2)
+			qdel(src)
+			return
 	..()
 
 /obj/item/firealarm_frame/proc/try_build(turf/on_wall)
